Key Execution demo
As we have not fully developed the database or communication interface
between systems, i have provided a simpl GUI with several buttons that
will execute the displayed command on the targeted application window.
Instructions are included in the form.

As we progress forward with the other parts of the project, this service
will ultimately be made to wait until it is sent a command from the
android interface and execute said command immediately.